+++ packages/nouns-sdk/CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] 2025-08-15
+
+### Added
+
+- `@nouns/sdk/treasury` actions:
+  - `readNounsTreasuryBalancesInUsd`
+  - `readNounsTreasuryBalancesInEth`
+- `@nouns/sdk/react/treasury` hooks:
+  - `useReadNounsTreasuryBalancesInUsd`
+  - `useReadNounsTreasuryBalancesInEth`
+
+### Removed
+
+- CJS exports. They were included by mistake, since the package depends on wagmi v2+ which is ESM only
+
 ## [1.1.0] 2025-08-08
 
 ### Added
@@ -60,6 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Started maintaining a changelog from this version onward.
 
 [Unreleased]: https://github.com/nounsDAO/nouns-monorepo/tree/master/packages/nouns-sdk
+[1.2.0]: https://github.com/nounsDAO/nouns-monorepo/tree/fb183939d/packages/nouns-sdk
 [1.1.0]: https://github.com/nounsDAO/nouns-monorepo/tree/967341a4b/packages/nouns-sdk
 [1.0.0]: https://github.com/nounsDAO/nouns-monorepo/tree/6e0b43054/packages/nouns-sdk
 [0.4.0]: https://github.com/nounsDAO/nouns-monorepo/tree/6e75b03a5/packages/nouns-sdk